backend/requirements.txt
flask
flask-cors
python-docx
pymupdf